refactor(pullcount): 精简抽数计算扫描逻辑
- 将抽数计算 Go 实现拆分为 action、OCR、仓库扫描、券配置和结果计算模块，降低 action.go 体积
- 收窄仓库标题 OCR 的 pipeline 期望，只让寻访券/情报书相关候选进入 Go 处理
- 删除 CurrentPageItems、PendingQuantity、UnknownNames 和疑似券名猜测逻辑，改为只按 vouchers.json 的显式配置累计
- 更新单测覆盖券分类、cell 状态记录和滚动探针判停逻辑

验证：
- go test ./pullcount
- go test ./...
- python tools/validate_schema.py --resource-dirs assets/resource --exclude-dirs assets/resource/gamedata assets/resource/image assets/resource/model --task-dirs assets/tasks
